Fix launch failure in sleep mode (#4)
The app fails to launch if the device is in sleep mode and the
request for vsync fails (tdm_client_vblank_wait returning error
as HW vblank is not available). To prevent it, we need to call
tdm_client_vblank_set_enable_fake(vblank, enable_fake) by setting
enable_fake == 1.